Relocating visualisation machine to stand alone module
